I've noted a verified state (commit 4d42b28) where the checks pass.
This branch points to commit 4d42b28, where:
- The Pipfile is configured with the necessary dependencies.
- Linting (ruff) passes.
- Type checking (mypy) passes.
- Pytest passes (with a PYTHONPATH=. workaround for mediawiki_agent).

No new file changes are introduced in this commit itself beyond what exists in 4d42b28; this branch is primarily to bookmark this verified state.
