+++ .jules/setup
@@ -1,0 +1,8 @@
+curl -fsSL https://pyenv.run | bash
+export PYENV_ROOT="$HOME/.pyenv"
+[[ -d $PYENV_ROOT/bin ]] && export PATH="$PYENV_ROOT/bin:$PATH"
+eval "$(pyenv init - bash)"
+# eval "$(pyenv virtualenv-init -)" ?
+pip install --user pipenv
+PATH=/home/jules/.local/bin:$PATH
+pipenv install
