+++ .jules/test
@@ -1,0 +1,3 @@
+pipenv run ./lint.sh && \
+    pipenv run mypy .  && \
+    pipenv run pytest
