+++ lint.sh
@@ -1,2 +1,3 @@
+#!/bin/sh
 ruff format .
 ruff check --fix .
