+++ pyenv_installer.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+#
+# Usage: curl https://pyenv.run | bash
+#
+# For more info, visit: https://github.com/pyenv/pyenv-installer
+#
+index_main() {
+    set -e
+    curl -s -S -L https://raw.githubusercontent.com/pyenv/pyenv-installer/master/bin/pyenv-installer | bash
+}
+
+index_main
